requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# PDF processing and OCR
pytesseract==0.3.10
pdf2image==1.17.0
PyPDF2==3.0.1
Pillow==10.1.0

# Vector database and embeddings
faiss-cpu==1.7.4
numpy==1.24.3
sentence-transformers==2.2.2
scikit-learn==1.3.0

# LLM integrations
openai==1.3.0
langchain==0.0.350
langchain-community==0.0.1
requests==2.31.0

# Text processing
regex==2023.10.3

# Environment and configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Logging and monitoring
loguru==0.7.2

# UI and visualization
plotly==5.17.0
pandas==2.1.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
flake8==6.1.0